reset blender addon state after each attempt to generate model
The blender add on can get in a state where it's stuck in processing, as the state doesn't always get reset after every attempt to generate a model. Currently it depends on a happy path, which won't always be the case and results in the addon becoming unsable and requires a restart of blender or the plugin to work again.